New structure and rewrite of some parts (#298)
Hey! 👋 

This PR change the structure of the documentation.
It uses the new structure that has been discussed and accepted a long time ago. This structure is way more easy to follow for newcomers or actual users.

Also, I started a complete rewrite of the documentation. I have added in this PR some parts that have already been rewritten.

**NOTE: Missing number are intentional**